requirements.txt
asgiref==3.8.1
attrs==25.1.0
boto3==1.36.9
botocore==1.36.9
Django==5.1.5
django-debug-toolbar==5.0.1
django-storages==1.14.4
django-unfold==0.46.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
uWSGI==2.0.28
